+++ reed-solomon-novelpoly/src/field/faster8/mod.rs
@@ -1,0 +1,9 @@
+#![cfg(target_feature = "avx")]
+
+#[cfg(feature = "f256")]
+pub mod f256;
+#[cfg(feature = "f256")]
+pub use self::f256::*;
+
+pub mod f2e16;
+pub use self::f2e16::*;
