+++ reed-solomon-novelpoly/src/field/mod.rs
@@ -8,3 +8,6 @@ mod gen;
 pub mod f256;
 
 pub mod f2e16;
+
+#[cfg(target_feature = "avx")]
+pub mod faster8;
